0. Что такое переменная. [#3610]

+++ src/main/java/ru/job4j/Variables.java
@@ -4,7 +4,7 @@ public class Variables {
     public static void main(String[] args) {
         int age = 18;
         System.out.println(age);
-        float length = (float) 78.0332;
+        float length = 78.0332F;
         System.out.println(length);
         boolean baby = true;
         System.out.println(baby);
